# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "CampaignOut": {
        "properties": {
          "assets_count": {
            "title": "Assets Count",
            "type": "integer"
          },
          "clips_approved": {
            "title": "Clips Approved",
            "type": "integer"
          },
          "clips_published": {
            "title": "Clips Published",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "source_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Id"
          },
          "source_instructions": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Instructions"
          },
          "source_metadata": {
            "additionalProperties": true,
            "title": "Source Metadata",
            "type": "object"
          },
          "source_provider": {
            "title": "Source Provider",
            "type": "string"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "spec": {
            "additionalProperties": true,
            "title": "Spec",
            "type": "object"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "status",
          "source_provider",
          "source_metadata",
          "spec",
          "assets_count",
          "clips_approved",
          "clips_published",
          "created_at",
          "updated_at"
        ],
        "title": "CampaignOut",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Backend coordinator for the clipping pipeline (VPS side). All scheduled work runs from OpenClaw automations.",
    "title": "Clipping API",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/campaigns": {
      "get": {
        "operationId": "list_all_campaigns_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "source_provider",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source Provider"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CampaignOut"
                  },
                  "title": "Response List All Campaigns Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List All",
        "tags": [
          "campaigns"
        ]
      },
      "post": {
        "operationId": "create_campaigns_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CampaignCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create",
        "tags": [
          "campaigns"
        ]
      }
    },
    "/campaigns/": {
      "get": {
        "operationId": "list_all_campaigns__get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "source_provider",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source Provider"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CampaignOut"
                  },
                  "title": "Response List All Campaigns  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List All",
        "tags": [
          "campaigns"
        ]
      },
      "post": {
        "operationId": "create_campaigns__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CampaignCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create",
        "tags": [
          "campaigns"
        ]
      }
    },
...
    "/campaigns/status-machine": {
      "get": {
        "description": "Manual transitions allowed from each status (single source of truth:\napp/services/campaign_transitions.py). Forward steps belong to the ticks.",
        "operationId": "status_machine_campaigns_status_machine_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusMachineOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Status Machine",
        "tags": [
          "campaigns"
        ]
      }
    },
    "/campaigns/{campaign_id}": {
      "delete": {
        "description": "Hard delete, only for `archived` campaigns without active jobs.\nCascades to assets, candidates, clips and clip publications.",
        "operationId": "delete_campaigns__campaign_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "campaign_id",
            "required": true,
            "schema": {
              "title": "Campaign Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete",
        "tags": [
          "campaigns"
        ]
      },
      "get": {
        "operationId": "get_one_campaigns__campaign_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "campaign_id",
            "required": true,
            "schema": {
              "title": "Campaign Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get One",
        "tags": [
          "campaigns"
        ]
      },
      "patch": {
        "description": "Partial update (see CampaignUpdate). 400 unknown status, 409 duplicate\nname / forbidden transition / non-editable field, 422 validation.",
        "operationId": "update_campaigns__campaign_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "campaign_id",
            "required": true,
            "schema": {
              "title": "Campaign Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CampaignUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update",
        "tags": [
          "campaigns"
        ]
      }
    },
    "/campaigns/{campaign_id}/enqueue": {
      "post": {
        "description": "Enqueue the download job for a campaign.\n\nIMPORTANT (2026-09-11 race-fix):\n  Previously this endpoint enqueued download + transcribe + render in a\n  single atomic batch, sending source_url as the input path for both\n  transcribe and render. That caused the Worker to try to read a remote\n  URL it had not downloaded yet, producing hard failures.\n\n  Correct flow now:\n    1) This endpoint enqueues ONLY the download job.\n    2) on_download_completed (job_state_transitions.py) auto-creates the\n       transcribe job for the same asset, using asset.local_path (the\n       real on-disk file path) instead of source_url.\n    3) on_transcribe_completed triggers ClipSelectionAgent to produce\n       candidates. Render jobs are created later by candidate_lifecycle\n       once a candidate is approved (not pre-created here).\n\nIdempotent: existing pending/processing download jobs for the same\ncampaign are left untouched. Returns the count of jobs created.\n\nThe campaign must be in status='scored' (pipeline v2, paso 3c). The\nfirst processable asset of the campaign becomes the source.",
        "operationId": "enqueue_pipeline_campaigns__campaign_id__enqueue_post",
        "parameters": [
          {
            "in": "path",
            "name": "campaign_id",
            "required": true,
            "schema": {
              "title": "Campaign Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Enqueue Pipeline Campaigns  Campaign Id  Enqueue Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Enqueue Pipeline",
        "tags": [
          "campaigns"
        ]
      }
    },
...
  }
}